updated files
I had trouble turning in this assignment as GitHub desktop would not recognize my changes I made to the files. Hopefully this is the best alternative way to do it. If something is incorrect or missing please let me know as I  have made sure everything works on my end!